- Caching now transparently forwards exceptions and keeps working when serializers throw exceptions - Improved database-layer, removed 2 bytes for cid in references - General performance improvements - Model-data not stored transactionally anymore. Store model (which is transactional) is guarding access to models - Improved updating plugins - Updated to dump methods of new query objects to avoid hash collisions for different queries

+++ PluginBase/src/org/bimserver/plugins/serializers/RemovableFileOutputStream.java
@@ -33,6 +33,10 @@ public abstract class RemovableFileOutputStream extends FileOutputStream {
 	}
 	
 	public void remove() throws IOException {
-		Files.delete(file);
+		if (Files.exists(file)) {
+			Files.delete(file);
+		}
 	}
+
+	public abstract void cancel() throws IOException;
 }
